@@ -4,115 +4,136 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pgsql;
 
-use Yiisoft\Db\Exception\InvalidArgumentException;
-use Yiisoft\Db\Expression\Expression;
+use InvalidArgumentException;
 use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Db\QueryBuilder\AbstractDMLQueryBuilder;
 
+use function array_map;
 use function implode;
-use function reset;
+use function str_ends_with;
+use function substr;
 
 /**
  * Implements a DML (Data Manipulation Language) SQL statements for PostgreSQL Server.
  */
 final class DMLQueryBuilder extends AbstractDMLQueryBuilder
 {
-    public function insertWithReturningPks(string $table, QueryInterface|array $columns, array &$params = []): string
+    public function insertReturningPks(string $table, array|QueryInterface $columns, array &$params = []): string
     {
-        $sql = $this->insert($table, $columns, $params);
-
+        $insertSql = $this->insert($table, $columns, $params);
         $tableSchema = $this->schema->getTableSchema($table);
+        $primaryKeys = $tableSchema?->getPrimaryKey() ?? [];
 
-        $returnColumns = [];
-        if ($tableSchema !== null) {
-            $returnColumns = $tableSchema->getPrimaryKey();
+        if (empty($primaryKeys)) {
+            return $insertSql;
         }
 
-        if (!empty($returnColumns)) {
-            $returning = [];
-            foreach ($returnColumns as $name) {
-                $returning[] = $this->quoter->quoteColumnName($name);
-            }
-            $sql .= ' RETURNING ' . implode(', ', $returning);
-        }
+        $primaryKeys = array_map($this->quoter->quoteColumnName(...), $primaryKeys);
 
-        return $sql;
+        return $insertSql . ' RETURNING ' . implode(', ', $primaryKeys);
     }
 
-    public function resetSequence(string $table, int|string $value = null): string
+    public function resetSequence(string $table, int|string|null $value = null): string
     {
         $tableSchema = $this->schema->getTableSchema($table);
-
-        if ($tableSchema !== null && ($sequence = $tableSchema->getSequenceName()) !== null) {
-            /**
-             * @link https://www.postgresql.org/docs/8.1/static/functions-sequence.html
-             */
-            $sequence = $this->quoter->quoteTableName($sequence);
-            $table = $this->quoter->quoteTableName($table);
-
-            if ($value === null) {
-                $pk = $tableSchema->getPrimaryKey();
-                $key = $this->quoter->quoteColumnName(reset($pk));
-                $value = "(SELECT COALESCE(MAX($key),0) FROM $table)+1";
-            }
-
-            return "SELECT SETVAL('$sequence',$value,false)";
-        }
 
         if ($tableSchema === null) {
             throw new InvalidArgumentException("Table not found: '$table'.");
         }
 
-        throw new InvalidArgumentException("There is not sequence associated with table '$table'.");
+        $sequence = $tableSchema->getSequenceName();
+
+        if ($sequence === null) {
+            throw new InvalidArgumentException("There is not sequence associated with table '$table'.");
+        }
+
+        /** @link https://www.postgresql.org/docs/8.1/static/functions-sequence.html */
+        $sequence = $this->quoter->quoteTableName($sequence);
+
+        if ($value === null) {
+            $table = $this->quoter->quoteTableName($table);
+            $key = $tableSchema->getPrimaryKey()[0];
+            $key = $this->quoter->quoteColumnName($key);
+            $value = "(SELECT COALESCE(MAX($key),0) FROM $table)+1";
+        }
+
+        return "SELECT SETVAL('$sequence',$value,false)";
     }
 
     public function upsert(
         string $table,
-        QueryInterface|array $insertColumns,
-        $updateColumns,
-        array &$params = []
+        array|QueryInterface $insertColumns,
+        array|bool $updateColumns = true,
+        array &$params = [],
     ): string {
         $insertSql = $this->insert($table, $insertColumns, $params);
 
-        /** @psalm-var array $uniqueNames */
-        [$uniqueNames, , $updateNames] = $this->prepareUpsertColumns(
-            $table,
-            $insertColumns,
-            $updateColumns,
-        );
+        [$uniqueNames, , $updateNames] = $this->prepareUpsertColumns($table, $insertColumns, $updateColumns);
 
         if (empty($uniqueNames)) {
             return $insertSql;
         }
 
-        if ($updateNames === []) {
+        if (empty($updateColumns) || $updateNames === []) {
             /** there are no columns to update */
-            $updateColumns = false;
-        }
-
-        if ($updateColumns === false) {
             return "$insertSql ON CONFLICT DO NOTHING";
         }
 
-        if ($updateColumns === true) {
-            $updateColumns = [];
-
-            /** @psalm-var string $name */
-            foreach ($updateNames as $name) {
-                $updateColumns[$name] = new Expression(
-                    'EXCLUDED.' . $this->quoter->quoteColumnName($name)
-                );
-            }
-        }
-
-        /**
-         * @psalm-var array $updateColumns
-         * @psalm-var string[] $uniqueNames
-         * @psalm-var string[] $updates
-         */
-        [$updates, $params] = $this->prepareUpdateSets($table, $updateColumns, $params);
+        $quotedUniqueNames = array_map($this->quoter->quoteColumnName(...), $uniqueNames);
+        $updates = $this->prepareUpsertSets($table, $updateColumns, $updateNames, $params);
 
         return $insertSql
-            . ' ON CONFLICT (' . implode(', ', $uniqueNames) . ') DO UPDATE SET ' . implode(', ', $updates);
+            . ' ON CONFLICT (' . implode(', ', $quotedUniqueNames) . ')'
+            . ' DO UPDATE SET ' . implode(', ', $updates);
+    }
+
+    public function upsertReturning(
+        string $table,
+        array|QueryInterface $insertColumns,
+        array|bool $updateColumns = true,
+        ?array $returnColumns = null,
+        array &$params = [],
+    ): string {
+        $upsertSql = $this->upsert($table, $insertColumns, $updateColumns, $params);
+
+        $returnColumns ??= $this->schema->getTableSchema($table)?->getColumnNames();
+
+        if (empty($returnColumns)) {
+            return $upsertSql;
+        }
+
+        $returnColumns = array_map($this->quoter->quoteColumnName(...), $returnColumns);
+
+        if (str_ends_with($upsertSql, ' ON CONFLICT DO NOTHING')) {
+            $tableName = $this->quoter->quoteTableName($table);
+            $dummyColumn = $this->getDummyColumn($table);
+
+            $uniqueNames = $this->prepareUpsertColumns($table, $insertColumns, $updateColumns)[0];
+            $quotedUniqueNames = array_map($this->quoter->quoteColumnName(...), $uniqueNames);
+
+            $upsertSql = substr($upsertSql, 0, -10)
+                . '(' . implode(', ', $quotedUniqueNames) . ')'
+                . " DO UPDATE SET $dummyColumn = $tableName.$dummyColumn";
+        }
+
+        return $upsertSql . ' RETURNING ' . implode(', ', $returnColumns);
+    }
+
+    private function getDummyColumn(string $table): string
+    {
+        /** @psalm-suppress PossiblyNullReference */
+        $columns = $this->schema->getTableSchema($table)->getColumns();
+
+        foreach ($columns as $column) {
+            if ($column->isPrimaryKey() || $column->isUnique()) {
+                continue;
+            }
+
+            /** @psalm-suppress PossiblyNullArgument */
+            return $this->quoter->quoteColumnName($column->getName());
+        }
+
+        /** @psalm-suppress PossiblyNullArgument, PossiblyFalseReference */
+        return $this->quoter->quoteColumnName(end($columns)->getName());
     }
 }
