@@ -6,7 +6,7 @@ namespace Yiisoft\Db\Pgsql\Tests;
 
 use Throwable;
 use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Exception\InvalidArgumentException;
+use InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidCallException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Pgsql\Tests\Support\TestTrait;
@@ -38,10 +38,10 @@ final class PdoConnectionTest extends CommonPdoConnectionTest
                 'name' => 'Some {{weird}} name',
                 'email' => 'test@example.com',
                 'address' => 'Some {{%weird}} address',
-            ]
+            ],
         )->execute();
 
-        $this->assertSame('4', $db->getLastInsertID('public.customer_id_seq'));
+        $this->assertSame('4', $db->getLastInsertId('public.customer_id_seq'));
 
         $db->close();
     }
@@ -63,7 +63,7 @@ final class PdoConnectionTest extends CommonPdoConnectionTest
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('PostgreSQL not support lastInsertId without sequence name.');
 
-        $db->getLastInsertID();
+        $db->getLastInsertId();
 
         $db->close();
     }
